requirements.txt
discord~=2.3.2
python-dotenv~=1.0.1
peewee~=3.17.8
textblob~=0.18.0.post0
translate~=3.6.1
openai~=1.55.0
scipy~=1.14.1
matplotlib~=3.9.2
numpy~=2.1.3
Ipython
pydub
discord.py~=2.4.0
requests~=2.32.3
ffmpeg-python
audioop-lts
nltk
beautifulsoup4